refactor(homepage): :recycle: rename the all component to start with `BestProduct` to match it's folder name
